revert(animations): remove old animation implementations
Breaks usage of old interfaces, but both patterns can't coexist, because they work on the same animation graph in a conflicting way
