+++ src/plugins/animations/src/components/animation_lookup.rs
@@ -1,21 +1,16 @@
 use bevy::prelude::*;
 use common::traits::{
-	animation::{AffectedAnimationBones2, AnimationKey, AnimationPath, PlayMode},
+	animation::{AffectedAnimationBones2, AnimationKey, PlayMode},
 	iterate::Iterate,
 };
 use std::collections::HashMap;
 
 #[derive(Component, Debug, PartialEq)]
-pub(crate) struct AnimationLookup<TAnimations = AnimationClips> {
-	pub(crate) animations: HashMap<AnimationPath, (TAnimations, AnimationMask)>,
-}
-
-#[derive(Component, Debug, PartialEq)]
-pub(crate) struct AnimationLookup2<TAnimationClips = AnimationClips> {
+pub(crate) struct AnimationLookup<TAnimationClips = AnimationClips> {
 	pub(crate) animations: HashMap<AnimationKey, AnimationLookupData<TAnimationClips>>,
 }
 
-impl Default for AnimationLookup2 {
+impl Default for AnimationLookup {
 	fn default() -> Self {
 		Self {
 			animations: HashMap::default(),
