+++ src/plugins/animations/src/lib.rs
@@ -11,13 +11,8 @@ use crate::{
 	components::{animation_lookup::AnimationClips, setup_animations::SetupAnimations},
 	system_params::animations::{AnimationsParamMut, override_animations::AnimationOverrideEvent},
 	systems::{
-		discover_animation_mask_bones::DiscoverMaskChains,
-		init_animation_components::InitAnimationComponents,
-		mask_animation_nodes::MaskAnimationNodes,
-		play_animation_clip2::PlayAnimationClip2,
-		remove_unused_animation_targets::RemoveUnusedAnimationTargets,
+		play_animation_clip::PlayAnimationClip2,
 		set_directional_animation_weights::SetDirectionalAnimationWeights,
-		set_directional_animation_weights2::SetDirectionalAnimationWeights2,
 	},
 };
 use bevy::{prelude::*, scene::SceneInstanceReady};
@@ -36,17 +31,13 @@ use common::{
 			RegisterAnimations,
 		},
 		handles_saving::HandlesSaving,
-		register_derived_component::RegisterDerivedComponent,
 		system_set_definition::SystemSetDefinition,
 		thread_safe::ThreadSafe,
 	},
 };
 use components::animation_dispatch::AnimationDispatch;
 use std::marker::PhantomData;
-use systems::{
-	dispatch_player_components::DispatchPlayerComponents,
-	play_animation_clip::PlayAnimationClip,
-};
+use systems::dispatch_player_components::DispatchPlayerComponents;
 
 pub struct AnimationsPlugin<TDependencies>(PhantomData<TDependencies>);
 
@@ -60,40 +51,23 @@ where
 }
 
 impl<TDependencies> RegisterAnimations for AnimationsPlugin<TDependencies> {
-	fn register_animations<TAgent>(app: &mut App)
+	fn register_animations<TAgent>(_: &mut App)
 	where
 		TAgent: Component + GetAnimationDefinitions + ConfigureNewAnimationDispatch,
 		for<'a> AnimationMask: From<&'a TAgent::TAnimationMask>,
 		for<'a> AffectedAnimationBones: From<&'a TAgent::TAnimationMask>,
 	{
-		app.register_derived_component::<TAgent, AnimationDispatch>()
-			.add_systems(
-				Update,
-				(
-					TAgent::init_animation_components::<AnimationGraph, AssetServer>,
-					TAgent::mask_animation_nodes,
-					TAgent::set_animation_mask_bones,
-					TAgent::remove_unused_animation_targets,
-				)
-					.chain()
-					.in_set(AnimationSystems),
-			);
 	}
 
-	fn register_movement_direction<TMovementDirection>(app: &mut App)
+	fn register_movement_direction<TMovementDirection>(_: &mut App)
 	where
 		TMovementDirection: Component + GetMovementDirection,
 	{
-		app.add_systems(
-			Update,
-			AnimationDispatch::set_directional_animation_weights::<TMovementDirection>
-				.in_set(AnimationSystems),
-		);
 	}
 }
 
 impl<TDependencies> HasAnimationsDispatch for AnimationsPlugin<TDependencies> {
-	type TAnimationDispatch = AnimationDispatch;
+	type TAnimationDispatch = AnimationDispatch<Animation>;
 }
 
 impl<TSavegame> Plugin for AnimationsPlugin<TSavegame>
@@ -101,8 +75,8 @@ where
 	TSavegame: ThreadSafe + HandlesSaving,
 {
 	fn build(&self, app: &mut App) {
-		type DispatchNew = AnimationDispatch<AnimationKey>;
-		TSavegame::register_savable_component::<DispatchNew>(app);
+		type Dispatch = AnimationDispatch<AnimationKey>;
+		TSavegame::register_savable_component::<Dispatch>(app);
 		app.add_observer(AnimationOverrideEvent::observe);
 		app.add_observer(SetupAnimations::insert_when::<SceneInstanceReady>);
 		app.add_systems(
@@ -110,29 +84,15 @@ where
 			(
 				SetupAnimations::init_masks::<AnimationGraphHandle, AnimationClips>,
 				SetupAnimations::init_bone_groups::<AnimationGraphHandle>,
-				SetupAnimations::remove_unused_animation_targets2::<AnimationGraphHandle>,
+				SetupAnimations::remove_unused_animation_targets::<AnimationGraphHandle>,
 				SetupAnimations::stop,
-				DispatchNew::track_in_self_and_children::<AnimationPlayer>().system(),
-				DispatchNew::track_in_self_and_children::<AnimationGraphHandle>().system(),
-				DispatchNew::distribute_player_components::<AnimationGraphHandle>,
-				DispatchNew::play_animation_clip_via2::<&mut AnimationPlayer>,
-				DispatchNew::set_directional_animation_weights2,
+				Dispatch::track_in_self_and_children::<AnimationPlayer>().system(),
+				Dispatch::track_in_self_and_children::<AnimationGraphHandle>().system(),
+				Dispatch::distribute_player_components::<AnimationGraphHandle>,
+				Dispatch::play_animation_clip::<&mut AnimationPlayer>,
+				Dispatch::set_directional_animation_weights,
 			)
 				.chain()
-				.in_set(AnimationSystems),
-		);
-
-		// FIXME: Remove when all consumers use new `HandlesAnimations` interface
-		type DispatchOld = AnimationDispatch<Animation>;
-		TSavegame::register_savable_component::<DispatchOld>(app);
-		app.add_systems(
-			Update,
-			(
-				DispatchOld::play_animation_clip_via::<&mut AnimationPlayer>,
-				DispatchOld::track_in_self_and_children::<AnimationPlayer>().system(),
-				DispatchOld::track_in_self_and_children::<AnimationGraphHandle>().system(),
-				DispatchOld::distribute_player_components::<AnimationGraphHandle>,
-			)
 				.in_set(AnimationSystems),
 		);
 	}
