+++ src/plugins/animations/src/systems.rs
@@ -1,10 +1,4 @@
-pub(crate) mod discover_animation_mask_bones;
 pub(crate) mod dispatch_player_components;
-pub(crate) mod init_animation_components;
-pub(crate) mod mask_animation_nodes;
 pub(crate) mod play_animation_clip;
-pub(crate) mod play_animation_clip2;
-pub(crate) mod remove_unused_animation_targets;
 pub(crate) mod set_directional_animation_weights;
-pub(crate) mod set_directional_animation_weights2;
 pub(crate) mod setup_animations;
