--- src/plugins/animations/src/systems/play_animation_clip2.rs
@@ -1,1334 +1,0 @@
-use crate::{
-	components::{
-		animation_dispatch::AnimationDispatch,
-		animation_lookup::{AnimationLookup2, AnimationLookupData},
-	},
-	traits::{
-		AnimationPlayers,
-		GetActiveAnimations,
-		IsPlaying,
-		RepeatAnimation,
-		ReplayAnimation,
-		StopAnimation,
-		asset_server::animation_graph::GetNodeMut,
-	},
-};
-use bevy::{ecs::query::QueryData, prelude::*};
-use common::traits::{
-	animation::{AnimationKey, AnimationPriority, PlayMode},
-	iterate::Iterate,
-	thread_safe::ThreadSafe,
-	wrap_handle::{GetHandle, WrapHandle},
-};
-use std::collections::HashSet;
-
-impl<TDispatch> PlayAnimationClip2 for TDispatch where
-	TDispatch: Component + AnimationPlayers + GetActiveAnimations<AnimationKey>
-{
-}
-
-pub(crate) trait PlayAnimationClip2
-where
-	Self: Component + AnimationPlayers + GetActiveAnimations<AnimationKey> + Sized,
-{
-	#[allow(clippy::type_complexity)]
-	fn play_animation_clip_via2<TAnimationPlayer>(
-		players: Query<TAnimationPlayer>,
-		dispatchers: Query<
-			(
-				&AnimationDispatch<AnimationKey>,
-				&AnimationLookup2,
-				&AnimationGraphHandle,
-			),
-			Changed<AnimationDispatch<AnimationKey>>,
-		>,
-		graphs: ResMut<Assets<AnimationGraph>>,
-	) where
-		TAnimationPlayer: QueryData
-			+ for<'a> QueryData<Item<'a>: IsPlaying<AnimationNodeIndex>>
-			+ for<'a> QueryData<Item<'a>: ReplayAnimation<AnimationNodeIndex>>
-			+ for<'a> QueryData<Item<'a>: RepeatAnimation<AnimationNodeIndex>>
-			+ for<'a> QueryData<Item<'a>: StopAnimation<AnimationNodeIndex>>,
-	{
-		play_animation_clip_via2(players, dispatchers, graphs)
-	}
-}
-
-#[allow(clippy::type_complexity)]
-fn play_animation_clip_via2<TAnimationPlayer, TDispatch, TGraph, TAnimations>(
-	mut players: Query<TAnimationPlayer>,
-	agents: Query<
-		(
-			&TDispatch,
-			&AnimationLookup2<TAnimations>,
-			&TGraph::TComponent,
-		),
-		Changed<TDispatch>,
-	>,
-	mut graphs: ResMut<Assets<TGraph>>,
-) where
-	TAnimationPlayer: QueryData,
-	TGraph: Asset + GetNodeMut + WrapHandle,
-	TDispatch: Component + AnimationPlayers + GetActiveAnimations<AnimationKey>,
-	for<'a> TAnimations: ThreadSafe + Iterate<'a, TItem = &'a AnimationNodeIndex>,
-	for<'a> TAnimationPlayer::Item<'a>: IsPlaying<AnimationNodeIndex>
-		+ ReplayAnimation<AnimationNodeIndex>
-		+ RepeatAnimation<AnimationNodeIndex>
-		+ StopAnimation<AnimationNodeIndex>,
-{
-	for (dispatcher, lookup, graph_component) in &agents {
-		for entity in dispatcher.animation_players() {
-			let Ok(mut player) = players.get_mut(entity) else {
-				continue;
-			};
-			let Some(graph) = graphs.get_mut(graph_component.get_handle()) else {
-				continue;
-			};
-			let active_animations = play_active(graph, &mut player, lookup, dispatcher);
-			let inactive = |data: &AnimationLookupData<TAnimations>| {
-				let clips = &data.animation_clips;
-				if clips.iterate().any(|clip| active_animations.contains(clip)) {
-					return None;
-				}
-				Some(clips.iterate().copied().collect::<Vec<_>>())
-			};
-			let inactive_animations = lookup.animations.values().filter_map(inactive).flatten();
-			stop(player, graph, inactive_animations);
-		}
-	}
-}
-
-fn play_active<TPlayer, TDispatcher, TGraph, TAnimations>(
-	graph: &mut TGraph,
-	player: &mut TPlayer,
-	lookup: &AnimationLookup2<TAnimations>,
-	dispatcher: &TDispatcher,
-) -> HashSet<AnimationNodeIndex>
-where
-	TPlayer: IsPlaying<AnimationNodeIndex>
-		+ ReplayAnimation<AnimationNodeIndex>
-		+ RepeatAnimation<AnimationNodeIndex>,
-	TDispatcher: GetActiveAnimations<AnimationKey>,
-	TGraph: Asset + GetNodeMut,
-	for<'a> TAnimations: Iterate<'a, TItem = &'a AnimationNodeIndex>,
-{
-	let mut higher_priority_mask = 0;
-	let mut active_animations = HashSet::default();
-
-	for priority in AnimationPriority::ordered_descending() {
-		let blocked_by_higher_priority = higher_priority_mask;
-
-		for key in dispatcher.get_active_animations(priority) {
-			let Some(animation_data) = lookup.animations.get(key) else {
-				continue;
-			};
-
-			for id in animation_data.animation_clips.iterate() {
-				let Some(animation_node) = graph.get_node_mut(*id) else {
-					continue;
-				};
-
-				active_animations.insert(*id);
-				animation_node.remove_mask(animation_data.mask);
-				animation_node.add_mask(blocked_by_higher_priority);
-				add(&mut higher_priority_mask, animation_data.mask);
-
-				if player.is_playing(*id) {
-					continue;
-				}
-
-				match animation_data.play_mode {
-					PlayMode::Repeat => player.repeat(*id),
-					PlayMode::Replay => player.replay(*id),
-				}
-			}
-		}
-	}
-
-	active_animations
-}
-
-fn stop<TPlayer, TGraph>(
-	mut player: TPlayer,
-	graph: &mut TGraph,
-	animations: impl Iterator<Item = AnimationNodeIndex>,
-) where
-	TPlayer: StopAnimation<AnimationNodeIndex>,
-	TGraph: GetNodeMut,
-{
-	for animation_id in animations {
-		if let Some(node) = graph.get_node_mut(animation_id) {
-			node.add_mask(AnimationMask::MAX);
-		}
-		player.stop_animation(animation_id);
-	}
-}
-
-fn add(dst: &mut AnimationMask, src: AnimationMask) {
-	*dst |= src;
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-	use crate::test_tools::leak_iterator;
-	use common::{
-		tools::action_key::slot::SlotKey,
-		traits::animation::{AffectedAnimationBones2, BoneName},
-	};
-	use macros::NestedMocks;
-	use mockall::{mock, predicate::eq};
-	use std::{
-		collections::{HashMap, VecDeque},
-		ops::DerefMut,
-		slice::Iter,
-	};
-	use testing::{NestedMocks, SingleThreadedApp, new_handle};
-
-	#[derive(Component)]
-	struct _Agent;
-
-	#[derive(Component, NestedMocks)]
-	struct _AnimationDispatch {
-		mock: Mock_AnimationDispatch,
-	}
-
-	impl AnimationPlayers for _AnimationDispatch {
-		type TIter = _Iter;
-
-		fn animation_players(&self) -> Self::TIter {
-			self.mock.animation_players()
-		}
-	}
-
-	impl GetActiveAnimations<AnimationKey> for _AnimationDispatch {
-		type TIter<'a>
-			= Iter<'a, AnimationKey>
-		where
-			Self: 'a,
-			AnimationKey: 'a;
-
-		fn get_active_animations<TPriority>(&self, priority: TPriority) -> Self::TIter<'_>
-		where
-			TPriority: Into<AnimationPriority> + 'static,
-		{
-			self.mock.get_active_animations(priority)
-		}
-	}
-
-	mock! {
-		_AnimationDispatch {}
-		impl AnimationPlayers for _AnimationDispatch {
-			type TIter = _Iter;
-
-			fn animation_players(&self) -> _Iter;
-		}
-		impl GetActiveAnimations<AnimationKey> for _AnimationDispatch {
-			type TIter<'a>
-				= Iter<'a, AnimationKey>
-			where
-				Self: 'a,
-				AnimationKey: 'a;
-
-			fn get_active_animations<TPriority>(&self, priority: TPriority) -> Iter<'static, AnimationKey>
-			where
-				TPriority: Into<AnimationPriority> + 'static;
-		}
-
-	}
-
-	#[derive(Clone, Default)]
-	struct _Animations(Vec<AnimationNodeIndex>);
-
-	impl<'a> Iterate<'a> for _Animations {
-		type TItem = &'a AnimationNodeIndex;
-		type TIter = Iter<'a, AnimationNodeIndex>;
-
-		fn iterate(&'a self) -> Iter<'a, AnimationNodeIndex> {
-			self.0.iter()
-		}
-	}
-
-	impl From<&[AnimationNodeIndex]> for _Animations {
-		fn from(animations: &[AnimationNodeIndex]) -> Self {
-			let animations = Vec::from_iter(animations.iter().copied());
-			Self(animations)
-		}
-	}
-
-	impl From<&Vec<AnimationNodeIndex>> for _Animations {
-		fn from(animations: &Vec<AnimationNodeIndex>) -> Self {
-			let animations = Vec::from_iter(animations.iter().copied());
-			Self(animations)
-		}
-	}
-
-	#[derive(Asset, TypePath, Default)]
-	struct _Graph {
-		nodes: HashMap<usize, AnimationGraphNode>,
-	}
-
-	impl WrapHandle for _Graph {
-		type TComponent = _GraphComponent;
-
-		fn wrap_handle(handle: Handle<Self>) -> Self::TComponent {
-			_GraphComponent(handle)
-		}
-	}
-
-	impl GetNodeMut for _Graph {
-		fn get_node_mut(
-			&mut self,
-			animation: AnimationNodeIndex,
-		) -> Option<&'_ mut AnimationGraphNode> {
-			self.nodes.get_mut(&animation.index())
-		}
-	}
-
-	impl<const N: usize> From<[(usize, AnimationGraphNode); N]> for _Graph {
-		fn from(nodes: [(usize, AnimationGraphNode); N]) -> Self {
-			Self {
-				nodes: HashMap::from(nodes),
-			}
-		}
-	}
-
-	#[derive(Component)]
-	struct _GraphComponent(Handle<_Graph>);
-
-	impl GetHandle for _GraphComponent {
-		type TAsset = _Graph;
-
-		fn get_handle(&self) -> &Handle<Self::TAsset> {
-			&self.0
-		}
-	}
-
-	#[derive(Clone)]
-	struct _Iter(VecDeque<Entity>);
-
-	impl<const N: usize> From<[Entity; N]> for _Iter {
-		fn from(value: [Entity; N]) -> Self {
-			_Iter(VecDeque::from(value))
-		}
-	}
-
-	impl Iterator for _Iter {
-		type Item = Entity;
-
-		fn next(&mut self) -> Option<Self::Item> {
-			self.0.pop_front()
-		}
-	}
-
-	#[derive(Component, NestedMocks)]
-	struct _AnimationPlayer {
-		mock: Mock_AnimationPlayer,
-	}
-
-	impl Default for _AnimationPlayer {
-		fn default() -> Self {
-			let mut mock = Mock_AnimationPlayer::default();
-			mock.expect_replay().return_const(());
-			mock.expect_repeat().return_const(());
-			mock.expect_stop_animation().return_const(());
-			mock.expect_is_playing().return_const(false);
-			Self { mock }
-		}
-	}
-
-	impl IsPlaying<AnimationNodeIndex> for Mut<'_, _AnimationPlayer> {
-		fn is_playing(&self, index: AnimationNodeIndex) -> bool {
-			self.mock.is_playing(index)
-		}
-	}
-
-	impl ReplayAnimation<AnimationNodeIndex> for Mut<'_, _AnimationPlayer> {
-		fn replay(&mut self, index: AnimationNodeIndex) {
-			self.mock.replay(index)
-		}
-	}
-
-	impl RepeatAnimation<AnimationNodeIndex> for Mut<'_, _AnimationPlayer> {
-		fn repeat(&mut self, index: AnimationNodeIndex) {
-			self.mock.repeat(index)
-		}
-	}
-
-	impl StopAnimation<AnimationNodeIndex> for Mut<'_, _AnimationPlayer> {
-		fn stop_animation(&mut self, index: AnimationNodeIndex) {
-			self.mock.stop_animation(index)
-		}
-	}
-
-	mock! {
-		_AnimationPlayer {}
-		impl ReplayAnimation<AnimationNodeIndex> for _AnimationPlayer {
-			fn replay(&mut self, index: AnimationNodeIndex);
-		}
-		impl RepeatAnimation<AnimationNodeIndex> for _AnimationPlayer {
-			fn repeat(&mut self, index: AnimationNodeIndex);
-		}
-		impl StopAnimation<AnimationNodeIndex> for _AnimationPlayer{
-			fn stop_animation(&mut self, index: AnimationNodeIndex);
-		}
-		impl IsPlaying<AnimationNodeIndex> for _AnimationPlayer {
-			fn is_playing(&self, index: AnimationNodeIndex) -> bool;
-		}
-	}
-
-	macro_rules! binary_str {
-		($a:expr) => {{
-			let values = $a
-				.into_iter()
-				.map(|v| format!("{v:b}"))
-				.collect::<Vec<_>>()
-				.join(", ");
-			format!("[{values}]")
-		}};
-	}
-
-	macro_rules! setup {
-		($animations:expr, $graph_handle:expr) => {
-			setup($animations, $graph_handle, AnimationMask::default())
-		};
-		($animations:expr, $graph_handle:expr, $initial_mask:expr) => {
-			setup($animations, $graph_handle, $initial_mask)
-		};
-	}
-
-	fn setup(
-		lookup: &AnimationLookup2<_Animations>,
-		graph_handle: &Handle<_Graph>,
-		initial_mask: AnimationMask,
-	) -> App {
-		let mut app = App::new().single_threaded(Update);
-		let mut graphs = Assets::default();
-		let mut graph = _Graph::default();
-
-		for data in lookup.animations.values() {
-			for animation in data.animation_clips.iterate() {
-				graph.nodes.insert(
-					animation.index(),
-					AnimationGraphNode {
-						mask: initial_mask,
-						..default()
-					},
-				);
-			}
-		}
-
-		graphs.insert(graph_handle, graph);
-		app.insert_resource(graphs);
-		app.add_systems(
-			Update,
-			play_animation_clip_via2::<
-				&mut _AnimationPlayer,
-				_AnimationDispatch,
-				_Graph,
-				_Animations,
-			>,
-		);
-
-		app
-	}
-
-	#[test]
-	fn repeat_animation() {
-		let handle = new_handle();
-		let indices = vec![
-			AnimationNodeIndex::new(1),
-			AnimationNodeIndex::new(2),
-			AnimationNodeIndex::new(3),
-		];
-		let lookup = AnimationLookup2 {
-			animations: HashMap::from([(
-				AnimationKey::Walk,
-				AnimationLookupData {
-					animation_clips: _Animations::from(&indices),
-					play_mode: PlayMode::Repeat,
-					..default()
-				},
-			)]),
-		};
-		let mut app = setup!(&lookup, &handle);
-		let animation_player = app
-			.world_mut()
-			.spawn(_AnimationPlayer::new().with_mock(assert_repeat(indices)))
-			.id();
-		app.world_mut().spawn((
-			_AnimationDispatch::new().with_mock(|mock| {
-				mock.expect_animation_players()
-					.return_const(_Iter::from([animation_player]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::High))
-					.return_const(leak_iterator(vec![AnimationKey::Walk]));
-				mock.expect_get_active_animations::<AnimationPriority>()
-					.return_const(leak_iterator(vec![]));
-			}),
-			lookup,
-			_GraphComponent(handle),
-		));
-
-		app.update();
-
-		fn assert_repeat(indices: Vec<AnimationNodeIndex>) -> impl Fn(&mut Mock_AnimationPlayer) {
-			move |mock| {
-				mock.expect_is_playing().return_const(false);
-				mock.expect_replay().never().return_const(());
-				for index in indices.clone() {
-					mock.expect_repeat()
-						.times(1)
-						.with(eq(index))
-						.return_const(());
-				}
-				mock.expect_stop_animation().never().return_const(());
-			}
-		}
-	}
-
-	#[test]
-	fn replay_animation() {
-		let handle = new_handle();
-		let indices = vec![
-			AnimationNodeIndex::new(1),
-			AnimationNodeIndex::new(2),
-			AnimationNodeIndex::new(3),
-		];
-		let lookup = AnimationLookup2 {
-			animations: HashMap::from([(
-				AnimationKey::Walk,
-				AnimationLookupData {
-					animation_clips: _Animations::from(&indices),
-					play_mode: PlayMode::Replay,
-					..default()
-				},
-			)]),
-		};
-		let mut app = setup!(&lookup, &handle);
-		let animation_player = app
-			.world_mut()
-			.spawn(_AnimationPlayer::new().with_mock(assert_replay(indices)))
-			.id();
-		app.world_mut().spawn((
-			_AnimationDispatch::new().with_mock(|mock| {
-				mock.expect_animation_players()
-					.return_const(_Iter::from([animation_player]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::High))
-					.return_const(leak_iterator(vec![AnimationKey::Walk]));
-				mock.expect_get_active_animations::<AnimationPriority>()
-					.return_const(leak_iterator(vec![]));
-			}),
-			lookup,
-			_GraphComponent(handle),
-		));
-
-		app.update();
-
-		fn assert_replay(indices: Vec<AnimationNodeIndex>) -> impl Fn(&mut Mock_AnimationPlayer) {
-			move |mock| {
-				mock.expect_is_playing().return_const(false);
-				mock.expect_repeat().never().return_const(());
-				for index in indices.clone() {
-					mock.expect_replay()
-						.times(1)
-						.with(eq(index))
-						.return_const(());
-				}
-				mock.expect_stop_animation().never().return_const(());
-			}
-		}
-	}
-
-	#[test]
-	fn play_all_animations() {
-		let handle = new_handle();
-		let indices = vec![
-			AnimationNodeIndex::new(1),
-			AnimationNodeIndex::new(2),
-			AnimationNodeIndex::new(3),
-			AnimationNodeIndex::new(4),
-			AnimationNodeIndex::new(5),
-			AnimationNodeIndex::new(6),
-			AnimationNodeIndex::new(7),
-			AnimationNodeIndex::new(8),
-			AnimationNodeIndex::new(9),
-			AnimationNodeIndex::new(10),
-			AnimationNodeIndex::new(11),
-			AnimationNodeIndex::new(12),
-		];
-		let lookup = AnimationLookup2 {
-			animations: HashMap::from([
-				(
-					AnimationKey::Skill(SlotKey(11)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[0..=1]),
-						play_mode: PlayMode::Replay,
-						mask: 0,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(12)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[2..=3]),
-						play_mode: PlayMode::Replay,
-						mask: 0,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(21)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[4..=5]),
-						play_mode: PlayMode::Replay,
-						mask: 0,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(22)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[6..=7]),
-						play_mode: PlayMode::Replay,
-						mask: 0,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(31)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[8..=9]),
-						play_mode: PlayMode::Replay,
-						mask: 0,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(32)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[10..=11]),
-						play_mode: PlayMode::Replay,
-						mask: 0,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-			]),
-		};
-		let mut app = setup!(&lookup, &handle);
-		let animation_player = app
-			.world_mut()
-			.spawn(_AnimationPlayer::new().with_mock(assert_repeat(indices)))
-			.id();
-		app.world_mut().spawn((
-			_AnimationDispatch::new().with_mock(|mock: &mut Mock_AnimationDispatch| {
-				mock.expect_animation_players()
-					.return_const(_Iter::from([animation_player]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::High))
-					.return_const(leak_iterator(vec![
-						AnimationKey::Skill(SlotKey(11)),
-						AnimationKey::Skill(SlotKey(12)),
-					]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::Medium))
-					.return_const(leak_iterator(vec![
-						AnimationKey::Skill(SlotKey(21)),
-						AnimationKey::Skill(SlotKey(22)),
-					]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::Low))
-					.return_const(leak_iterator(vec![
-						AnimationKey::Skill(SlotKey(31)),
-						AnimationKey::Skill(SlotKey(32)),
-					]));
-			}),
-			lookup,
-			_GraphComponent(handle),
-		));
-
-		app.update();
-
-		fn assert_repeat(indices: Vec<AnimationNodeIndex>) -> impl Fn(&mut Mock_AnimationPlayer) {
-			move |mock| {
-				mock.expect_is_playing().return_const(false);
-				mock.expect_repeat().never().return_const(());
-				for index in indices.clone() {
-					mock.expect_replay()
-						.times(1)
-						.with(eq(index))
-						.return_const(());
-				}
-				mock.expect_stop_animation().return_const(());
-			}
-		}
-	}
-
-	#[test]
-	fn do_not_play_animation_which_is_playing() {
-		let handle = new_handle();
-		let indices = vec![
-			AnimationNodeIndex::new(1),
-			AnimationNodeIndex::new(2),
-			AnimationNodeIndex::new(3),
-		];
-		let lookup = AnimationLookup2 {
-			animations: HashMap::from([(
-				AnimationKey::Walk,
-				AnimationLookupData {
-					animation_clips: _Animations::from(&indices),
-					play_mode: PlayMode::Repeat,
-					..default()
-				},
-			)]),
-		};
-		let mut app = setup!(&lookup, &handle);
-		let animation_player = app
-			.world_mut()
-			.spawn(_AnimationPlayer::new().with_mock(assert_not_playing(indices)))
-			.id();
-		app.world_mut().spawn((
-			_AnimationDispatch::new().with_mock(|mock: &mut Mock_AnimationDispatch| {
-				mock.expect_animation_players()
-					.return_const(_Iter::from([animation_player]));
-
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::High))
-					.return_const(leak_iterator(vec![AnimationKey::Walk]));
-				mock.expect_get_active_animations::<AnimationPriority>()
-					.return_const(leak_iterator(vec![]));
-			}),
-			lookup,
-			_GraphComponent(handle),
-		));
-
-		app.update();
-
-		fn assert_not_playing(
-			indices: Vec<AnimationNodeIndex>,
-		) -> impl Fn(&mut Mock_AnimationPlayer) {
-			move |mock| {
-				for i in indices.clone() {
-					mock.expect_is_playing()
-						.times(1)
-						.with(eq(i))
-						.return_const(true);
-				}
-				mock.expect_replay().never().return_const(());
-				mock.expect_repeat().never().return_const(());
-				mock.expect_stop_animation().never().return_const(());
-			}
-		}
-	}
-
-	#[test]
-	fn stop_playing_animation_not_returned_in_dispatcher() {
-		let handle = new_handle();
-		let indices = vec![
-			AnimationNodeIndex::new(1),
-			AnimationNodeIndex::new(2),
-			AnimationNodeIndex::new(3),
-		];
-		let lookup = AnimationLookup2 {
-			animations: HashMap::from([(
-				AnimationKey::Walk,
-				AnimationLookupData {
-					animation_clips: _Animations::from(&indices),
-					play_mode: PlayMode::Repeat,
-					..default()
-				},
-			)]),
-		};
-		let mut app = setup!(&lookup, &handle);
-		let animation_player = app
-			.world_mut()
-			.spawn(_AnimationPlayer::new().with_mock(assert_stop(indices)))
-			.id();
-		app.world_mut().spawn((
-			_AnimationDispatch::new().with_mock(|mock: &mut Mock_AnimationDispatch| {
-				mock.expect_animation_players()
-					.return_const(_Iter::from([animation_player]));
-
-				mock.expect_get_active_animations::<AnimationPriority>()
-					.return_const(leak_iterator(vec![]));
-			}),
-			lookup,
-			_GraphComponent(handle),
-		));
-
-		app.update();
-
-		fn assert_stop(indices: Vec<AnimationNodeIndex>) -> impl Fn(&mut Mock_AnimationPlayer) {
-			move |mock| {
-				mock.expect_is_playing().return_const(false);
-				mock.expect_replay().return_const(());
-				mock.expect_repeat().return_const(());
-				for i in indices.clone() {
-					mock.expect_stop_animation()
-						.times(1)
-						.with(eq(i))
-						.return_const(());
-				}
-			}
-		}
-	}
-
-	#[test]
-	fn play_animation_only_once() {
-		let handle = new_handle();
-		let lookup = AnimationLookup2 {
-			animations: HashMap::from([(
-				AnimationKey::Walk,
-				AnimationLookupData {
-					animation_clips: _Animations::from(&vec![AnimationNodeIndex::new(1)]),
-					play_mode: PlayMode::Repeat,
-					..default()
-				},
-			)]),
-		};
-		let mut app = setup!(&lookup, &handle);
-		let animation_player = app
-			.world_mut()
-			.spawn(_AnimationPlayer::new().with_mock(assert_repeat_once))
-			.id();
-		app.world_mut().spawn((
-			_AnimationDispatch::new().with_mock(|mock: &mut Mock_AnimationDispatch| {
-				mock.expect_animation_players()
-					.return_const(_Iter::from([animation_player]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::High))
-					.return_const(leak_iterator(vec![AnimationKey::Walk]));
-				mock.expect_get_active_animations::<AnimationPriority>()
-					.return_const(leak_iterator(vec![]));
-			}),
-			lookup,
-			_GraphComponent(handle),
-		));
-
-		app.update();
-		app.update();
-
-		fn assert_repeat_once(mock: &mut Mock_AnimationPlayer) {
-			mock.expect_is_playing().return_const(false);
-			mock.expect_replay().never().return_const(());
-			mock.expect_repeat().times(1).return_const(());
-			mock.expect_stop_animation().never().return_const(());
-		}
-	}
-
-	#[test]
-	fn play_animation_again_after_dispatcher_mutably_dereferenced() {
-		let handle = new_handle();
-		let lookup = AnimationLookup2 {
-			animations: HashMap::from([(
-				AnimationKey::Walk,
-				AnimationLookupData {
-					animation_clips: _Animations::from(&vec![AnimationNodeIndex::new(1)]),
-					play_mode: PlayMode::Repeat,
-					..default()
-				},
-			)]),
-		};
-		let mut app = setup!(&lookup, &handle);
-		let animation_player = app
-			.world_mut()
-			.spawn(_AnimationPlayer::new().with_mock(assert_repeat_twice))
-			.id();
-		let agent = app
-			.world_mut()
-			.spawn((
-				_AnimationDispatch::new().with_mock(|mock: &mut Mock_AnimationDispatch| {
-					mock.expect_animation_players()
-						.return_const(_Iter::from([animation_player]));
-					mock.expect_get_active_animations()
-						.with(eq(AnimationPriority::High))
-						.return_const(leak_iterator(vec![AnimationKey::Walk]));
-					mock.expect_get_active_animations::<AnimationPriority>()
-						.return_const(leak_iterator(vec![]));
-				}),
-				lookup,
-				_GraphComponent(handle),
-			))
-			.id();
-
-		app.update();
-		app.world_mut()
-			.entity_mut(agent)
-			.get_mut::<_AnimationDispatch>()
-			.unwrap()
-			.deref_mut();
-		app.update();
-
-		fn assert_repeat_twice(mock: &mut Mock_AnimationPlayer) {
-			mock.expect_is_playing().return_const(false);
-			mock.expect_replay().never().return_const(());
-			mock.expect_repeat().times(2).return_const(());
-			mock.expect_stop_animation().never().return_const(());
-		}
-	}
-
-	#[test]
-	fn mask_depending_on_priority() {
-		let handle = new_handle();
-		let indices = [
-			AnimationNodeIndex::new(1),
-			AnimationNodeIndex::new(2),
-			AnimationNodeIndex::new(3),
-			AnimationNodeIndex::new(4),
-			AnimationNodeIndex::new(5),
-			AnimationNodeIndex::new(6),
-			AnimationNodeIndex::new(7),
-			AnimationNodeIndex::new(8),
-			AnimationNodeIndex::new(9),
-			AnimationNodeIndex::new(10),
-			AnimationNodeIndex::new(11),
-			AnimationNodeIndex::new(12),
-		];
-		let lookup = AnimationLookup2 {
-			animations: HashMap::from([
-				(
-					AnimationKey::Skill(SlotKey(11)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[0..=1]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 0,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(12)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[2..=3]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 1,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(21)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[4..=5]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 2,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(22)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[6..=7]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 3,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(31)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[8..=9]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 4,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(32)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[10..=11]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 5,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-			]),
-		};
-		let mut app = setup!(&lookup, &handle);
-		let animation_player = app.world_mut().spawn(_AnimationPlayer::default()).id();
-		app.world_mut().spawn((
-			_AnimationDispatch::new().with_mock(|mock: &mut Mock_AnimationDispatch| {
-				mock.expect_animation_players()
-					.return_const(_Iter::from([animation_player]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::High))
-					.return_const(leak_iterator(vec![
-						AnimationKey::Skill(SlotKey(11)),
-						AnimationKey::Skill(SlotKey(12)),
-					]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::Medium))
-					.return_const(leak_iterator(vec![
-						AnimationKey::Skill(SlotKey(21)),
-						AnimationKey::Skill(SlotKey(22)),
-					]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::Low))
-					.return_const(leak_iterator(vec![
-						AnimationKey::Skill(SlotKey(31)),
-						AnimationKey::Skill(SlotKey(32)),
-					]));
-			}),
-			lookup,
-			_GraphComponent(handle.clone()),
-		));
-
-		app.update();
-
-		let graph = app
-			.world()
-			.resource::<Assets<_Graph>>()
-			.get(&handle)
-			.unwrap();
-		let masks = &indices
-			.iter()
-			.map(|i| graph.nodes.get(&i.index()).unwrap().mask)
-			.collect::<Vec<_>>();
-		// each priority has 2 assets each with 2 animations
-		//   -> 4 animations masked by higher priority mask per priority
-		let expected = &std::iter::repeat_n(0b000000, 4)
-			.chain(std::iter::repeat_n(0b000011, 4))
-			.chain(std::iter::repeat_n(0b001111, 4))
-			.collect::<Vec<_>>();
-		assert_eq!(
-			expected,
-			masks,
-			"\n  left bits: {}\n right bits: {}",
-			binary_str!(expected),
-			binary_str!(masks)
-		);
-	}
-
-	#[test]
-	fn unmask_depending_on_priority() {
-		let handle = new_handle();
-		let indices = [
-			AnimationNodeIndex::new(1),
-			AnimationNodeIndex::new(2),
-			AnimationNodeIndex::new(3),
-			AnimationNodeIndex::new(4),
-			AnimationNodeIndex::new(5),
-			AnimationNodeIndex::new(6),
-			AnimationNodeIndex::new(7),
-			AnimationNodeIndex::new(8),
-			AnimationNodeIndex::new(9),
-			AnimationNodeIndex::new(10),
-			AnimationNodeIndex::new(11),
-			AnimationNodeIndex::new(12),
-		];
-		let lookup = AnimationLookup2 {
-			animations: HashMap::from([
-				(
-					AnimationKey::Skill(SlotKey(11)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[0..=1]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 0,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(12)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[2..=3]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 1,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(21)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[4..=5]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 2,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(22)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[6..=7]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 3,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(31)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[8..=9]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 4,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(32)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[10..=11]),
-						play_mode: PlayMode::Replay,
-						mask: 1 << 5,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-			]),
-		};
-		let initial_mask = 0b111111;
-		let mut app = setup!(&lookup, &handle, initial_mask);
-		let animation_player = app.world_mut().spawn(_AnimationPlayer::default()).id();
-		app.world_mut().spawn((
-			_AnimationDispatch::new().with_mock(|mock: &mut Mock_AnimationDispatch| {
-				mock.expect_animation_players()
-					.return_const(_Iter::from([animation_player]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::High))
-					.return_const(leak_iterator(vec![
-						AnimationKey::Skill(SlotKey(11)),
-						AnimationKey::Skill(SlotKey(12)),
-					]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::Medium))
-					.return_const(leak_iterator(vec![
-						AnimationKey::Skill(SlotKey(21)),
-						AnimationKey::Skill(SlotKey(22)),
-					]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::Low))
-					.return_const(leak_iterator(vec![
-						AnimationKey::Skill(SlotKey(31)),
-						AnimationKey::Skill(SlotKey(32)),
-					]));
-			}),
-			lookup,
-			_GraphComponent(handle.clone()),
-		));
-
-		app.update();
-
-		let graph = app
-			.world()
-			.resource::<Assets<_Graph>>()
-			.get(&handle)
-			.unwrap();
-		let masks = &indices
-			.iter()
-			.map(|i| graph.nodes.get(&i.index()).unwrap().mask)
-			.collect::<Vec<_>>();
-		// each asset has 2 animations
-		//   -> 2 animations masked per asset
-		let expected = &std::iter::repeat_n(0b111110, 2)
-			.chain(std::iter::repeat_n(0b111101, 2))
-			.chain(std::iter::repeat_n(0b111011, 2))
-			.chain(std::iter::repeat_n(0b110111, 2))
-			.chain(std::iter::repeat_n(0b101111, 2))
-			.chain(std::iter::repeat_n(0b011111, 2))
-			.collect::<Vec<_>>();
-		assert_eq!(
-			expected,
-			masks,
-			"\n  left bits: {}\n right bits: {}",
-			binary_str!(expected),
-			binary_str!(masks)
-		);
-	}
-
-	#[test]
-	fn set_mask_for_already_playing_animation() {
-		let handle = new_handle();
-		let indices = [
-			AnimationNodeIndex::new(1),
-			AnimationNodeIndex::new(2),
-			AnimationNodeIndex::new(3),
-			AnimationNodeIndex::new(4),
-		];
-		let lookup = AnimationLookup2 {
-			animations: HashMap::from([
-				(
-					AnimationKey::Skill(SlotKey(1)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[0..=1]),
-						play_mode: PlayMode::Replay,
-						mask: 0b000001,
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-				(
-					AnimationKey::Skill(SlotKey(2)),
-					AnimationLookupData {
-						animation_clips: _Animations::from(&indices[2..=3]),
-						play_mode: PlayMode::Replay,
-						mask: 0b000111, // wants to play on high masks (..11)
-						bones: AffectedAnimationBones2 {
-							from_root: BoneName::from(""),
-							..default()
-						},
-					},
-				),
-			]),
-		};
-		let initial_mask = 0b111111;
-		let mut app = setup!(&lookup, &handle, initial_mask);
-		let animation_player = app
-			.world_mut()
-			.spawn(_AnimationPlayer::new().with_mock(|mock| {
-				mock.expect_is_playing()
-					.with(eq(indices[2]))
-					.return_const(true);
-				mock.expect_is_playing()
-					.with(eq(indices[3]))
-					.return_const(true);
-				mock.expect_is_playing().return_const(false);
-				mock.expect_repeat().return_const(());
-				mock.expect_replay().return_const(());
-				mock.expect_stop_animation().return_const(());
-			}))
-			.id();
-		app.world_mut().spawn((
-			_AnimationDispatch::new().with_mock(|mock: &mut Mock_AnimationDispatch| {
-				mock.expect_animation_players()
-					.return_const(_Iter::from([animation_player]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::High))
-					.return_const(leak_iterator(vec![AnimationKey::Skill(SlotKey(1))]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::Medium))
-					.return_const(leak_iterator(vec![AnimationKey::Skill(SlotKey(2))]));
-				mock.expect_get_active_animations()
-					.with(eq(AnimationPriority::Low))
-					.return_const(leak_iterator(vec![]));
-			}),
-			lookup,
-			_GraphComponent(handle.clone()),
-		));
-
-		app.update();
-
-		let graph = app
-			.world()
-			.resource::<Assets<_Graph>>()
-			.get(&handle)
-			.unwrap();
-		let masks = &indices
-			.iter()
-			.map(|i| graph.nodes.get(&i.index()).unwrap().mask)
-			.collect::<Vec<_>>();
-		// each asset has 2 animations
-		//   -> 2 animations masked per asset
-		let expected = &std::iter::repeat_n(0b111110, 2)
-			.chain(std::iter::repeat_n(0b111001, 2))
-			.collect::<Vec<_>>();
-		assert_eq!(
-			expected,
-			masks,
-			"\n  left bits: {}\n right bits: {}",
-			binary_str!(expected),
-			binary_str!(masks)
-		);
-	}
-
-	#[test]
-	fn completely_mask_animations_not_returned_by_dispatcher() {
-		let handle = new_handle();
-		let indices = vec![AnimationNodeIndex::new(1), AnimationNodeIndex::new(2)];
-		let lookup = AnimationLookup2 {
-			animations: HashMap::from([(
-				AnimationKey::Walk,
-				AnimationLookupData {
-					animation_clips: _Animations::from(&indices),
-					play_mode: PlayMode::Repeat,
-					..default()
-				},
-			)]),
-		};
-		let mut app = setup!(&lookup, &handle);
-		let animation_player = app.world_mut().spawn(_AnimationPlayer::default()).id();
-		app.world_mut().spawn((
-			_AnimationDispatch::new().with_mock(|mock: &mut Mock_AnimationDispatch| {
-				mock.expect_animation_players()
-					.return_const(_Iter::from([animation_player]));
-
-				mock.expect_get_active_animations::<AnimationPriority>()
-					.return_const(leak_iterator(vec![]));
-			}),
-			lookup,
-			_GraphComponent(handle.clone()),
-		));
-
-		app.update();
-
-		let graph = app
-			.world()
-			.resource::<Assets<_Graph>>()
-			.get(&handle)
-			.unwrap();
-		let masks = &indices
-			.iter()
-			.map(|i| graph.nodes.get(&i.index()).unwrap().mask)
-			.collect::<Vec<_>>();
-		let expected = &vec![AnimationMask::MAX, AnimationMask::MAX];
-		assert_eq!(
-			expected,
-			masks,
-			"\n  left bits: {}\n right bits: {}",
-			binary_str!(expected),
-			binary_str!(masks)
-		);
-	}
-}
