+++ src/plugins/animations/src/systems/setup_animations.rs
@@ -1,4 +1,4 @@
 pub(crate) mod init_bone_groups;
 pub(crate) mod init_masks;
-pub(crate) mod remove_unused_animation_targets2;
+pub(crate) mod remove_unused_animation_targets;
 pub(crate) mod stop;
